+++ src/popup/app/services/filter-events.ts
